fix[lembrete]: corrigido tipo de dados do id

+++ src/main/java/br/edu/ufape/agiota/negocio/basica/Lembrete.java
@@ -1,20 +1,16 @@
 package br.edu.ufape.agiota.negocio.basica;
 
-import java.util.Date;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.*;
 import lombok.Data;
+
+import java.util.Date;
 
 @Data
 @Entity
 public class Lembrete {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private Date data;
     private String texto;
